‼️Refactor binary module ‼️Require new auth ‼️
Remove some orphaned code.

On Linux remove detection of XDG_DATA_HOME, so it would fail in
situations where in some shell it is present and in another one not, so
it would double the places where config file and binary collected.

For Linux users would require to auth and download binary file again.